found bug in dpw3: SOLVE TOMORROW!
it originates from convert and convert_sp actually, but only manisfests
itself in dpw3. The problem is that I don't convert the elements with a ==b or d == e
correctly. I loop over a < b or a + S_ab <= b but I don't set the elements a ==b equal to zero
when S_ab == 1.